Extract EverymenSelectCard into its own file, repainting nothing (#2327)
The pure move, so the repaint that follows is reviewable on its own. The
component, its `CAST` constant and the `EverymenSigil` import leave
`FactionSelectCard.tsx`, which keeps the dispatcher, the shared types and
`LEGACY_SLUG`; `factions/everymen.ts` registers the new file as a plain
`lazyArchetype` rather than reaching into the dispatcher for a named export.

Proven byte-identical: `renderToStaticMarkup` over all six `state` x `members`
combinations, against the same component read from `origin/main`, 22,435 bytes
each and `cmp` clean.
